# api/openapi.json
{
  "components": {
...
    "headers": {
...
      "RequestID": {
        "description": "Correlation ID assigned to the request.",
        "example": "b618a705-2928-4f41-ab1f-c2fd95622887",
        "schema": {
          "maxLength": 128,
          "type": "string"
        }
      }
    },
    "parameters": {
...
      "TransactionID": {
        "description": "Financial transaction UUID.",
        "example": "33333333-3333-4333-8333-333333333333",
        "in": "path",
        "name": "transactionId",
        "required": true,
        "schema": {
          "format": "uuid",
          "type": "string"
        }
      },
...
      "UserID": {
        "description": "UUID asserted by a trusted upstream identity component. It determines\nwallet ownership and the idempotency namespace. It is not a substitute\nfor authentication on an untrusted public deployment.\n",
        "example": "44444444-4444-4444-8444-444444444444",
        "in": "header",
        "name": "X-User-ID",
        "required": true,
        "schema": {
          "format": "uuid",
          "type": "string"
        }
      },
...
    },
...
    "responses": {
      "BadRequest": {
        "$ref": "#/components/responses/Problem400"
      },
...
      "Forbidden": {
        "$ref": "#/components/responses/Problem403"
      },
...
      "InternalServerError": {
        "$ref": "#/components/responses/Problem500"
      },
      "MethodNotAllowed": {
        "content": {
          "application/problem+json": {
            "example": {
              "code": "METHOD_NOT_ALLOWED",
              "detail": "This resource does not support the requested HTTP method.",
              "instance": "/v1/wallets",
              "requestId": "b618a705-2928-4f41-ab1f-c2fd95622887",
              "status": 405,
              "title": "Method not allowed",
              "type": "https://wallet.example/problems/METHOD_NOT_ALLOWED"
            },
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "The resource does not support the requested HTTP method.",
        "headers": {
          "Allow": {
            "$ref": "#/components/headers/Allow"
          },
          "X-Request-ID": {
            "$ref": "#/components/headers/RequestID"
          }
        }
      },
      "NotFound": {
        "$ref": "#/components/responses/Problem404"
      },
...
      "OperationOK": {
        "content": {
          "application/json": {
            "examples": {
              "transfer": {
                "$ref": "#/components/examples/TransferOperation"
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Operation"
            }
          }
        },
        "description": "Committed financial operation.",
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/RequestID"
          }
        }
      },
...
      "ServiceUnavailable": {
        "$ref": "#/components/responses/Problem503"
      },
...
    },
...
  },
  "info": {
    "contact": {
      "name": "Wallet Service Team"
    },
    "description": "REST API for creating single-currency wallets and performing deposits,\nwithdrawals, transfers, and transaction-history queries.\n\nFinancial mutations are atomic and idempotent. Amounts are decimal strings,\nnever JSON numbers. `X-User-ID` represents a trusted upstream principal in\nthis assignment; a public deployment must derive it from verified credentials.\n\nOnly successfully committed financial operations appear in history. Rejected\nrequests do not create financial transactions or ledger entries.\n",
    "license": {
      "name": "Proprietary"
    },
    "summary": "Atomic, owner-scoped wallet operations",
    "title": "Wallet Service API",
    "version": "1.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/docs": {
      "get": {
        "description": "Redirects to the canonical trailing-slash Swagger UI URL when documentation is enabled.",
        "operationId": "redirectToSwaggerUI",
        "responses": {
          "307": {
            "description": "Temporary redirect to `/docs/`.",
            "headers": {
              "Location": {
                "description": "Canonical Swagger UI path.",
                "example": "/docs/",
                "schema": {
                  "const": "/docs/",
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          }
        },
        "summary": "Redirect to Swagger UI",
        "tags": [
          "Documentation"
        ],
        "x-development-only": true
      }
    },
    "/docs/": {
      "get": {
        "description": "Interactive Swagger UI, available only when Swagger documentation is enabled.",
        "operationId": "swaggerUI",
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Swagger UI HTML and embedded assets."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          }
        },
        "summary": "Open Swagger UI",
        "tags": [
          "Documentation"
        ],
        "x-development-only": true
      }
    },
...
    "/openapi.json": {
      "get": {
        "description": "Generated from openapi.yaml and available only when Swagger documentation is enabled.",
        "operationId": "getOpenAPIGenerated",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Generated OpenAPI 3.1 JSON document.",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestID"
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          }
        },
        "summary": "Download the generated OpenAPI document",
        "tags": [
          "Documentation"
        ],
        "x-development-only": true
      }
    },
    "/openapi.yaml": {
      "get": {
        "description": "Available only when Swagger documentation is enabled.",
        "operationId": "getOpenAPISource",
        "responses": {
          "200": {
            "content": {
              "application/yaml": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "OpenAPI 3.1 YAML source.",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestID"
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          }
        },
        "summary": "Download the OpenAPI source document",
        "tags": [
          "Documentation"
        ],
        "x-development-only": true
      }
    },
    "/v1/transactions/{transactionId}": {
      "get": {
        "description": "Returns a committed operation to an owner of a participating wallet.",
        "operationId": "getTransaction",
        "responses": {
          "200": {
            "$ref": "#/components/responses/OperationOK"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "summary": "Get a financial transaction",
        "tags": [
          "Transactions"
        ]
      },
      "parameters": [
        {
          "$ref": "#/components/parameters/TransactionID"
        },
        {
          "$ref": "#/components/parameters/UserID"
        },
        {
          "$ref": "#/components/parameters/RequestID"
        }
      ]
    },
    "/v1/transfers": {
      "post": {
        "description": "Atomically debits the caller-owned source wallet and credits the\ndestination wallet. Both wallets must use the request currency. Wallets\nare locked in stable UUID order to avoid opposing-transfer deadlocks.\n",
        "operationId": "transfer",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserID"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/RequestID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "transferUSD": {
                  "summary": "Transfer USD between different wallets",
                  "value": {
                    "amount": "25.00",
                    "currency": "USD",
                    "destinationWalletId": "22222222-2222-4222-8222-222222222222",
                    "sourceWalletId": "11111111-1111-4111-8111-111111111111"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/TransferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OperationReplay"
          },
          "201": {
            "$ref": "#/components/responses/OperationCreated"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "summary": "Transfer funds between wallets",
        "tags": [
          "Operations"
        ]
      }
    },
    "/v1/wallets": {
      "post": {
        "description": "Creates a zero-balance wallet owned by the asserted user. An owner can\nhave at most one wallet. Replaying the same idempotency key and request\nreturns the original wallet with status 200.\n",
        "operationId": "createWallet",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserID"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/RequestID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "usdWallet": {
                  "summary": "Create a USD wallet",
                  "value": {
                    "currency": "USD"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/CreateWalletRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/WalletReplay"
          },
          "201": {
            "$ref": "#/components/responses/WalletCreated"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "summary": "Create a wallet",
        "tags": [
          "Wallets"
        ]
      }
    },
    "/v1/wallets/{walletId}": {
      "get": {
        "description": "Returns the wallet and current materialized balance to its owner.",
        "operationId": "getWallet",
        "responses": {
          "200": {
            "$ref": "#/components/responses/WalletOK"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "summary": "Get a wallet",
        "tags": [
          "Wallets"
        ]
      },
      "parameters": [
        {
          "$ref": "#/components/parameters/WalletID"
        },
        {
          "$ref": "#/components/parameters/UserID"
        },
        {
          "$ref": "#/components/parameters/RequestID"
        }
      ]
    },
    "/v1/wallets/{walletId}/deposits": {
      "parameters": [
        {
          "$ref": "#/components/parameters/WalletID"
        },
        {
          "$ref": "#/components/parameters/UserID"
        },
        {
          "$ref": "#/components/parameters/RequestID"
        }
      ],
      "post": {
        "description": "Atomically credits the owner's wallet and appends a transaction and ledger entry.",
        "operationId": "deposit",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Money"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OperationReplay"
          },
          "201": {
            "$ref": "#/components/responses/OperationCreated"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "summary": "Deposit funds",
        "tags": [
          "Operations"
        ]
      }
    },
    "/v1/wallets/{walletId}/transactions": {
      "get": {
        "description": "Returns committed ledger effects newest first. The opaque cursor selects\nrecords older than the previous page and must be used without modification.\n",
        "operationId": "listWalletTransactions",
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/TransactionTypeFilter"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/HistoryOK"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "summary": "List wallet transaction history",
        "tags": [
          "Transactions"
        ]
      },
      "parameters": [
        {
          "$ref": "#/components/parameters/WalletID"
        },
        {
          "$ref": "#/components/parameters/UserID"
        },
        {
          "$ref": "#/components/parameters/RequestID"
        }
      ]
    },
    "/v1/wallets/{walletId}/withdrawals": {
      "parameters": [
        {
          "$ref": "#/components/parameters/WalletID"
        },
        {
          "$ref": "#/components/parameters/UserID"
        },
        {
          "$ref": "#/components/parameters/RequestID"
        }
      ],
      "post": {
        "description": "Atomically debits the owner's wallet and appends history. The operation\nfails without changing state when the available balance is insufficient.\n",
        "operationId": "withdraw",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Money"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OperationReplay"
          },
          "201": {
            "$ref": "#/components/responses/OperationCreated"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "summary": "Withdraw funds",
        "tags": [
          "Operations"
        ]
      }
    }
  },
  "servers": [
    {
      "description": "Local development",
      "url": "http://localhost:8080"
    }
  ],
...
}